// src/api/api-course/models/ProjectDaysMapper.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project & Day API
 * An OpenAPI specification for project management, day templates, tasks, forms, media, and related user progress tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProjectDaysMapperId } from './ProjectDaysMapperId';
import {
    ProjectDaysMapperIdFromJSON,
    ProjectDaysMapperIdFromJSONTyped,
    ProjectDaysMapperIdToJSON,
    ProjectDaysMapperIdToJSONTyped,
} from './ProjectDaysMapperId';

/**
 * Entity for mapping days within a project
 * @export
 * @interface ProjectDaysMapper
 */
export interface ProjectDaysMapper {
    /**
     * 
     * @type {ProjectDaysMapperId}
     * @memberof ProjectDaysMapper
     */
    id: ProjectDaysMapperId;
    /**
     * Order or position within a list
     * @type {number}
     * @memberof ProjectDaysMapper
     */
    sortOrder: number;
}

/**
 * Check if a given object implements the ProjectDaysMapper interface.
 */
export function instanceOfProjectDaysMapper(value: object): value is ProjectDaysMapper {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('sortOrder' in value) || value['sortOrder'] === undefined) return false;
    return true;
}

export function ProjectDaysMapperFromJSON(json: any): ProjectDaysMapper {
    return ProjectDaysMapperFromJSONTyped(json, false);
}

export function ProjectDaysMapperFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectDaysMapper {
    if (json == null) {
        return json;
    }
    return {
        
        'id': ProjectDaysMapperIdFromJSON(json['id']),
        'sortOrder': json['sortOrder'],
    };
}

export function ProjectDaysMapperToJSON(json: any): ProjectDaysMapper {
    return ProjectDaysMapperToJSONTyped(json, false);
}

export function ProjectDaysMapperToJSONTyped(value?: ProjectDaysMapper | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': ProjectDaysMapperIdToJSON(value['id']),
        'sortOrder': value['sortOrder'],
    };
}